requirements.txt
playwright==1.38.0
pytest==7.4.2
python_dateutil==2.8.2
PyYAML==6.0.1
typer==0.9.0
typing_extensions==4.8.0
